MySQL support - first version
