+++ build.gradle
@@ -38,29 +38,32 @@ distTar {
 }
 
 applicationDistribution.from("src/support/sample") {
-
     into "./sample"
-
 }
 
 applicationDistribution.from("src/support/install") {
     applicationDistribution.from("src/support/sample") {
-
         into "./sample"
-
     }
     filter { String line ->
         line.startsWith('version=') ? "version=\"${version}\"" : line
     }
 }
 
-/**
- * This is a support task for testing contents of tar, run it from terminal
- *
- *     $ sh ./gradlew checkTar
- *
- */
-
+// * This is a support task for testing contents of tar, run it from terminal
+// *     $ sh ./gradlew checkTar
 task checkTar(type: Exec){
     commandLine 'sh', 'src/support/checkinstall', projectDir, buildDir, version
 }
+
+test {
+    exclude '**/integration/*'
+}
+
+task testMySQL (type: Test){
+    systemProperty "test.myslq.user", "testuser"
+    systemProperty "test.myslq.password", "test"
+    systemProperty "test.myslq.uri","jdbc:mysql://localhost:13306/"
+    systemProperty "test.myslq.database", "testdb"
+}
+
